Tony Visualizations and formatting
Created a visualization version of the data with cleaned up names for printing on plots, added options for descriptive data with distribution plots, added box plots of student scores by race, mother's education, and sex. Added density plots of variables. Formatted the document to have a table of contents and code folding.  Oh also added a quick ANOVA, but just for fun.